Restore AVS_VERSION define
AVS_VERSION had been changed to AVS_CLASSIC_VERSION many years ago
to allow the VersionNumber() function to continue to report '2.6'
even though the early versioning used for AviSynth+ was '0.1',
which caused problems when VersionNumber returned *that* instead.

When we changed to using 3.x version numbers, this ended up being
left behind on account of probably forgetting that was why the
AVS_CLASSIC_VERSION thing happened in the first place. It likely
should have been changed back at that time.